pygameinstall
Claire: This should be the beginning of installing pygame. It took me 30-60 minutes in class and outside of class to figure out how to important pygame in my script vs on the python terminal. But this part should open pygame for a split second as the rest hasn't been coded yet. 25.04.2025.