Scenario3start
Claire:  6 hours. This part took me a while. In this section I have been figuring out how to code Scenario 3, especially where in the actual code I should put it, some parts are in the section where I define scenario 3, others are in the main loop of the game. Figuring out the wrapped function also took a while but allowed me to make sure my text always fits inside the boxes. Something for later, will be making sure this code, as everything is on one text file, is readable and not too cumbersome. I still need to figure out the pathways from scenario3 scene  1 to scene 2 given the options that the use selects. This part also took a while because I had to debug it, figure out how to put writing within the boxes, why at some point all of the text was showing up on the start screen, why the game closed every time I selected scenario 3 instead of moving to scene 2 etc...