scenario3more
Claire: 6 hours. There aren't many changes here, a lot of it is debugging, which took a lot of time. I realized that I had two main loops. What I'm still trying to do here is make it so that when a choice is selected in scene 1, the screen moves to scene 2, I still need to figure out how to potentially make a new screen for the recommendations or make it so that after the recommendation pops up after the player selected a choice, that the player can't then go back and select a new option