scenario3scene3
Claire: 5 hours, I spent more time debugging, the code for scenario 3 now goes through all three scenes. I had some trouble with the timer at first and making sure the screen changed when a choice was selected. There are still some issues because the code doesnt 'lock in ' the user's answer but shows all of the recommendations so I still need to fix that for all scenes