DebuggingScenario1
Claire: 3hour30. I was debugging Scenario 1, figuring out how to get the options to fit into the buttons, to get the scene transitions to work, and removing excess code to make sure the code is readable. This was done through a lot of trial and error and trying to model Scenario 3